frontend/owdex/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.1
dnspython==2.3.0
exceptiongroup==1.1.0
Flask==2.2.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonlines==3.1.0
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pycparser==2.21
pymongo==4.3.3
pysolr==3.9.0
pytest==7.2.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
toml==0.10.2
tomli==2.0.1
url-normalize==1.4.3
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.3
